+++ app/(tabs)/siktemerker/components/useCalcForm.ts
